Next Patient Function
Implemented the patient class with the MainGUI. The priorities depend on your Input (High/Medium/Low) also your age and whether if you are already a patient or not.  changes to the request class to send the data to the PQueue. The button for after the test is functional and the missing person will go to a stack list which i am am working on right now.